+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* updated timers to use unit `rsyslog-housekeeping.service`
 * renamed installed service file from `weekly-rsyslog-housekeeping.service` to `rsyslog-housekeeping.service`
 
 ## [1.0.0](https://github.com/bazzline/rsyslog_mysql_housekeeping/tree/1.0.0) - released at 20210426