+++ CHANGELOG.md
@@ -27,8 +27,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+## [1.0.1](https://github.com/bazzline/rsyslog_mysql_housekeeping/tree/1.0.1) - released at 20210521
+
+### Changed
+
 * updated timers to use unit `rsyslog-housekeeping.service`
 * renamed installed service file from `weekly-rsyslog-housekeeping.service` to `rsyslog-housekeeping.service`
+* fixed systemd detection in install and uninstall scripts
 
 ## [1.0.0](https://github.com/bazzline/rsyslog_mysql_housekeeping/tree/1.0.0) - released at 20210426